ServiceTwoServer/requirements.txt
astroid>=2.3.3
certifi>=2019.9.11
chardet>=3.0.4
Click>=7.0
colorama>=0.4.1
etcd3>=0.8.1
Flask>=1.1.1
grpcio>=1.18.0
grpcio-tools>=1.18.0
grpc-micorservice-discover>=0.0.3
idna>=2.8
isort>=4.3.21
itsdangerous>=1.1.0
Jinja2>=2.10.3
lazy-object-proxy>=1.4.3
MarkupSafe>=1.1.1
mccabe>=0.6.1
pip>=19.0.3
protobuf>=3.10.0
pylint>=2.4.4
pyodbc>=4.0.27
requests>=2.22.0
setuptools>=40.8.0
simplejson>=3.16.0
six>=1.13.0
tenacity>=6.0.0
typed-ast>=1.4.0
urllib3>=1.25.7
Werkzeug>=0.16.0
wrapt>=1.11.2
